+++ lib/collection.dart
@@ -47,7 +47,7 @@ class Collection<T> implements Queryable<T>, StoreReference {
     PersistorSettings? persistorSettings,
     DependenciesBuilder<S>? dependenciesBuilder,
   }) {
-    final (parent, id) = _splitReferencePath(path);
+    final (parent, id) = splitReferencePath(path);
 
     return Collection<S>(
       parent,