+++ lib/persistor/file_persistor/file_data_store_config.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 import 'package:loon/loon.dart';
+import 'package:loon/src/store/store.dart';
 import 'package:loon/persistor/data_store.dart';
 import 'package:loon/persistor/data_store_resolver.dart';