+++ test/core/dependency_traversal_test.dart
@@ -1,6 +1,7 @@
 import 'package:fake_async/fake_async.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:loon/loon.dart';
+import 'package:loon/src/store/store.dart';
 
 ValueStore<BroadcastEvents> pendingEvents() =>
     ValueStore(Loon.inspect()['broadcastStore']['events']);